pyacmecapture: update code to support use of real device
Up to know code was developed using only the virtual device
VirtualIIOAcmeProbe, because real physical device access was
not yet implemented.
Now pyacmecapture can both access virtual and physical devices.

Note that with the new IIOAcmeProbe there is no more limitation to
8 slots per cape. This used to be hard-coded and was removed.
This code was successfully tested with 2 ACME capes mounted on a
single BBB.

Signed-off-by: Patrick Titiano <ptitiano@baylibre.com>